Update dice_rolling_simulator.py
Someone suggested me not to use while too much, so tried to create an alternative. 
I am a newbie and this is my first contribution!
A line in the previous code (The equivalent to which in the current code is line #26 said [10], which should have been [12], please fix it. Thanks.